requirements.txt
flask
flask_pymongo
Pillow
numpy
scikit-image
